test: further split scanner coverage tests
